Backend/requirements.txt
Django==3.2.25
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
djongo==1.3.6
pymongo==3.12.3
python-dotenv==1.0.1
boto3==1.34.122
django-storages==1.14.3
django-cors-headers==4.3.1
Pillow==10.3.0
